--- ExperienceApp/Screen - Destaques/View/PopUp/ProfileViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ProfileViewController.swift
-//  ExperienceApp
-//
-//  Created by Tamara Erlij on 21/05/20.
-//  Copyright © 2020 Zewu Chen. All rights reserved.
-//
-
-import UIKit
-
-class ProfileViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        DispatchQueue.main.asyncAfter(deadline: .now() + 3) {
-            self.dismiss(animated: true, completion: nil)
-        }
-    }
-
-}
