+++ ExperienceApp/Screen - Destaques/View/PopUp/View/Cells/ExpAdicionadaPopUp.swift
@@ -12,7 +12,6 @@ class ExpAdicionadaPopUp: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
     }
 
     @IBAction func btnConcluído(_ sender: UIButton) {
