+++ ExperienceApp/Screen - Destaques/View/PopUp/View/Cells/ExpCriadaPopUp.swift
@@ -12,7 +12,6 @@ class ExpCriadaPopUp: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
     }
 
     @IBAction func btnConcluído(_ sender: UIButton) {
