+++ ExperienceApp/Screen - Destaques/View/PopUp/View/Cells/HomeViewController.xib
@@ -57,7 +57,7 @@
                     </connections>
                 </button>
             </subviews>
-            <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+            <color key="backgroundColor" red="0.71764705880000002" green="0.30196078430000001" blue="0.30196078430000001" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
             <constraints>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="6Q2-FV-oNy" secondAttribute="trailing" constant="20" id="BfR-AW-zK0"/>
                 <constraint firstItem="6Q2-FV-oNy" firstAttribute="top" secondItem="Ir4-hc-9Xz" secondAttribute="bottom" constant="30" id="KpQ-AU-huV"/>
